Hudson wasn't properly streaming a large external build submission, which can result in OOME and unresponsiveness.
While investigating the sudden memory usage spike, I discovered that XMPP/MXParser doesn't have a capability of returning a long PCDATA as a series of events. Instead, it tries to buffer the whole thing and return it as a single string.
This obviously results in a horrible performance problem if the submission is large, as was the case.

Switching to StAX to prevent this problem. StAX API isn't a part of Java5, but we already bundle woodstox.

git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@27042 71c3de6d-444a-0410-be80-ed276b4c234a

Originally-Committed-As: 819ab89dda473c04ce621bc214fb9a2fd7092e63